Stop publishing test artifacts & internal modules
Isolates the consumable test artifacts for d2 and darkcluster into the
new modules d2-test-api and darkcluster-test-api, respectively. Also
discontinues publishing "internal" modules to Bintray, with these
modules determined as modules with no main source set and no known
consumers. Both these changes are backward-incompatible.
