+++ d2-test-api/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+  compile project(':d2')
+  compile externalDependency.testng
+}
