+++ d2-test-api/src/main/java/com/linkedin/d2/discovery/stores/zk/ZKTestUtil.java
@@ -14,14 +14,11 @@
    limitations under the License.
 */
 
-/**
- * $Id: $
- */
-
 package com.linkedin.d2.discovery.stores.zk;
 
 import org.apache.zookeeper.WatchedEvent;
 import org.apache.zookeeper.Watcher;
+import org.testng.Assert;
 
 import java.io.File;
 import java.io.IOException;
@@ -33,13 +30,9 @@ import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
-import static org.testng.Assert.fail;
-
 /**
  * @author Steven Ihde
- * @version $Revision: $
  */
-
 public class ZKTestUtil
 {
   private static final int TEMP_DIR_ATTEMPTS = 10;
@@ -75,7 +68,7 @@ public class ZKTestUtil
     }
     catch (IOException e)
     {
-      fail("unable to instantiate real zk server.");
+      Assert.fail("unable to instantiate real zk server.");
       e.printStackTrace();
     }
