+++ darkcluster-test-api/src/main/java/com/linkedin/darkcluster/MockClusterInfoProvider.java
