+++ darkcluster/build.gradle
@@ -5,5 +5,6 @@ dependencies {
   compile project(':restli-client')
   compile project(':pegasus-common')
   testCompile externalDependency.testng
+  testCompile project(':darkcluster-test-api')
   testCompile project(path: ':d2', configuration: 'testArtifacts')
 }
