+++ gradle.properties
@@ -1,4 +1,4 @@
-version=28.3.11
+version=29.0.0
 group=com.linkedin.pegasus
 org.gradle.configureondemand=true
 org.gradle.parallel=true
