+++ settings.gradle
@@ -12,7 +12,9 @@ include 'd2-benchmark'
 include 'd2-schemas'
 include 'd2-contrib'
 include 'd2-int-test'
+include 'd2-test-api'
 include 'darkcluster'
+include 'darkcluster-test-api'
 include 'degrader'
 include 'entity-stream'
 include 'li-jersey-uri'
